Refactor for reusing detalle publicacion view. Added navigation from anfitrion publicaciones list to detalle.
