+++ app/src/main/java/com/example/bookbnb/ui/busqueda/ReservaRealizadaFragment.kt
@@ -8,16 +8,18 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.lifecycle.ViewModelProvider
 import com.example.bookbnb.R
+import com.example.bookbnb.viewmodels.DetallePublicacionHuespedViewModel
+import com.example.bookbnb.viewmodels.DetallePublicacionHuespedViewModelFactory
 import com.example.bookbnb.viewmodels.DetallePublicacionViewModel
 import com.example.bookbnb.viewmodels.DetallePublicacionViewModelFactory
 
 class ReservaRealizadaFragment : Fragment() {
-    private val viewModel: DetallePublicacionViewModel by lazy {
+    private val viewModel: DetallePublicacionHuespedViewModel by lazy {
         val activity = requireNotNull(this.activity) {
             "You can only access the viewModel after onActivityCreated()"
         }
-        ViewModelProvider(activity, DetallePublicacionViewModelFactory(activity.application))
-            .get(DetallePublicacionViewModel::class.java)
+        ViewModelProvider(activity, DetallePublicacionHuespedViewModelFactory(activity.application))
+            .get(DetallePublicacionHuespedViewModel::class.java)
     }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
