+++ app/src/main/res/layout/fragment_detalle_publicacion_huesped.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:bind="http://schemas.android.com/apk/res-auto">
+    <data>
+        <variable
+            name="detallePublicacionViewModel"
+            type="com.example.bookbnb.viewmodels.DetallePublicacionHuespedViewModel" />
+    </data>
+        <ScrollView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <include android:id="@+id/detallePublicacion"
+                    layout="@layout/detalle_publicacion"
+                    bind:detallePublicacionViewModel="@{detallePublicacionViewModel}"/>
+
+                <Button
+                    android:id="@+id/reservarBtn"
+                    style="@style/PrimaryButton"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="@dimen/margin_normal"
+                    android:layout_marginBottom="@dimen/margin_normal"
+                    android:onClick="@{() -> detallePublicacionViewModel.onReservaButtonClick()}"
+                    android:text="@string/reservar_lbl"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toBottomOf="@+id/detallePublicacion" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </ScrollView>
+</layout>
