+++ app/src/main/res/layout/publicacion_item.xml
@@ -69,6 +69,8 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/titulo"
+            android:drawableStart="@drawable/ic_location_24"
+            android:gravity="center_vertical"
             tools:text="Paseo Colón 850" />
 
         <TextView
